requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.103.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
psycopg2==2.9.7
pyasn1==0.5.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
